requirements.txt
azure-ai-ml
azure-identity
matplotlib
seaborn
matplotlib
pandas
numpy
scikit-learn
joblib
